third_party/blink/web_tests/virtual/prefetch-serviceworker/external/wpt/speculation-rules/prefetch/tentative/service-worker/basic.sub.https_origin=same-site_sw=fetch-handler-to-fallback-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch with ServiceWorker (fetch-handler-to-fallback)
  assert_equals: prefetch requests should not be sent to the server. expected 0 but got 1
Harness: the test ran to completion.
